Create MainApp to reroute logic flow
