+++ build.gradle
@@ -1,25 +1,15 @@
 plugins {
-    id 'java'
     id 'application'
-    id 'com.github.johnrengelman.shadow' version '7.1.2'
-    id 'checkstyle'
-}
-
-application {
-    mainClass.set("shoai.Main")  // Custom for your project
+    id 'com.github.johnrengelman.shadow' version '7.1.2' // To create a fat JAR
 }
 
 repositories {
     mavenCentral()
 }
 
+
 dependencies {
-    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.10.0'
-    testImplementation 'junit:junit:4.13.1'
-    testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.10.0'
-
     String javaFxVersion = '17.0.7'
-
     implementation group: 'org.openjfx', name: 'javafx-base', version: javaFxVersion, classifier: 'win'
     implementation group: 'org.openjfx', name: 'javafx-base', version: javaFxVersion, classifier: 'mac'
     implementation group: 'org.openjfx', name: 'javafx-base', version: javaFxVersion, classifier: 'linux'
@@ -32,40 +22,17 @@ dependencies {
     implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'win'
     implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'mac'
     implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'linux'
+    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.10.0'
+    testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.10.0'
 }
 
-test {
-    useJUnitPlatform()
-
-    testLogging {
-        events "passed", "skipped", "failed"
-
-        showExceptions true
-        exceptionFormat "full"
-        showCauses true
-        showStackTraces true
-        showStandardStreams = false
-    }
+application {
+    mainClass.set('shoai.Main') // Update with your main class
 }
 
 shadowJar {
-    archiveFileName = 'shoai.jar'  // Custom for your project
-    archiveBaseName = "ShoAI"
-    archiveClassifier = null  // No classifier for the fat JAR
-}
-
-run {
-    standardInput = System.in
-    enableAssertions = true
-    jvmArgs = [
-            '--module-path', configurations.runtimeClasspath.asPath,
-    ]
-}
-
-checkstyle {
-    toolVersion = '10.2'
-}
-
-processResources {
-    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
+    archiveBaseName.set('shoai') // Your JAR name
+    archiveClassifier.set('')    // No classifier to make the JAR name 'shoai.jar'
+    archiveVersion.set('')       // No version in the name
+    mergeServiceFiles()
 }
